# Conflicts:
#	Video_Share_Django/开发日志.md

+++ Video_Share_Django/UserCommunication/views.py
@@ -18,14 +18,20 @@ def followuser(request):
     if request.method == 'POST':
         userID = request.POST.get('userID')
         followed_userID = request.POST.get('followedUserID')
-        user = UserInfo.objects.get(userID=userID)
-        followed_user = UserInfo.objects.get(userID=followed_userID)
-        UserConnection.objects.create(followerUser=user, followedUser=followed_user)
-        followed_user.FansNum = followed_user.FansNum + 1
-        followed_user.save()
-        user.ConcernsNum = user.ConcernsNum + 1
-        user.save()
-        return JsonResponse({'error': 0, 'msg': "关注成功"})
+        try:
+            user = UserInfo.objects.get(userID=userID)
+            followed_user = UserInfo.objects.get(userID=followed_userID)
+        except:
+            return JsonResponse({'error': 4002, 'msg': "关注用户或被关注用户不存在"})
+        if UserConnection.objects.filter(followerUser=user, followedUser=followed_user).exists():
+            return JsonResponse({'error': 4001, 'msg': "您已经关注过了"})
+        else:
+            UserConnection.objects.create(followerUser=user, followedUser=followed_user)
+            followed_user.FansNum = followed_user.FansNum + 1
+            followed_user.save()
+            user.ConcernsNum = user.ConcernsNum + 1
+            user.save()
+            return JsonResponse({'error': 0, 'msg': "关注成功"})
     else:
         return JsonResponse({'error': 2001, 'msg': "请求方式错误"})
 
@@ -35,14 +41,20 @@ def cancelfollow(request):
     if request.method == 'POST':
         userID = request.POST.get('userID')
         followed_userID = request.POST.get('followedUserID')
-        user = UserInfo.objects.get(userID=userID)
-        followed_user = UserInfo.objects.get(userID=followed_userID)
-        UserConnection.objects.get(followerUser=user, followedUser=followed_user).delete()
-        followed_user.FansNum = followed_user.FansNum - 1
-        followed_user.save()
-        user.ConcernsNum = user.ConcernsNum - 1
-        user.save()
-        return JsonResponse({'error': 0, 'msg': "取消关注成功"})
+        try:
+            user = UserInfo.objects.get(userID=userID)
+            followed_user = UserInfo.objects.get(userID=followed_userID)
+        except:
+            return JsonResponse({'error': 4002, 'msg': "关注用户或被关注用户不存在"})
+        if UserConnection.objects.filter(followerUser=user, followedUser=followed_user).exists():
+            UserConnection.objects.get(followerUser=user, followedUser=followed_user).delete()
+            followed_user.FansNum = followed_user.FansNum - 1
+            followed_user.save()
+            user.ConcernsNum = user.ConcernsNum - 1
+            user.save()
+            return JsonResponse({'error': 0, 'msg': "取消关注成功"})
+        else:
+            return JsonResponse({'error': 4001, 'msg': "您还没有关注过"})
     else:
         return JsonResponse({'error': 2001, 'msg': "请求方式错误"})
 
@@ -53,8 +65,11 @@ def sendletter(request):
         letter_userID = request.POST.get('letterUserID')
         lettered_userID = request.POST.get('letteredUserID')
         letter_text = request.POST.get('letterText')
-        letter_user = UserInfo.objects.get(userID=letter_userID)
-        lettered_user = UserInfo.objects.get(userID=lettered_userID)
+        try:
+            letter_user = UserInfo.objects.get(userID=letter_userID)
+            lettered_user = UserInfo.objects.get(userID=lettered_userID)
+        except:
+            return JsonResponse({'error': 4002, 'msg': "发送用户或接收用户不存在"})
         if letter_user.userLimit == 1:
             UserLetter.objects.create(letterUser=letter_user, letteredUser=lettered_user, letterText=letter_text)
             return JsonResponse({'error': 0, 'msg': "私信已发送"})
@@ -115,7 +130,7 @@ def enterhomepage(request):
                 'userPortrait': user.userAvatar,
                 'userInformation': user.userInformation,
             }
-            fanslist.append(user)
+            fanslist.append(fans_item)
 
         concernslist = []
         for user in list(UserConnection.objects.filter(followerUser=entered_user)):
