+++ Video_Share_Django/VideoInteraction/views.py
@@ -16,9 +16,17 @@ def like(request):
     if request.method == 'POST':
         userID = request.POST.get('userID')
         videoID = request.POST.get('videoID')
-        like_user = UserInfo.objects.get(userID=userID)
-        video = VideoInfo.objects.get(videoID=videoID)
+        try:
+            like_user = UserInfo.objects.get(userID=userID)
+        except:
+            return JsonResponse({'error': 4002, 'msg': "用户不存在"})
+        try:
+            video = VideoInfo.objects.get(videoID=videoID)
+        except:
+            return JsonResponse({'error': 4003, 'msg': "视频不存在"})
         liked_user = video.videoUpUser
+        if LikeRecord.objects.filter(likeUser=like_user, likedVideo=video).exists():
+            return JsonResponse({'error': 4001, 'msg': "已点赞"})
         LikeRecord.objects.create(likeUser=like_user, likedUser=liked_user, likeVideo=video)
         video.videoLikeNum = video.videoLikeNum + 1
         liked_user.TotalLikeNum = liked_user.TotalLikeNum + 1 # 更新用户点赞总数
@@ -34,15 +42,24 @@ def cancellike(request):
     if request.method == 'POST':
         userID = request.POST.get('userID')
         videoID = request.POST.get('videoID')
-        like_user = UserInfo.objects.get(userID=userID)
-        video = VideoInfo.objects.get(videoID=videoID)
+        try:
+            like_user = UserInfo.objects.get(userID=userID)
+        except:
+            return JsonResponse({'error': 4002, 'msg': "用户不存在"})
+        try:
+            video = VideoInfo.objects.get(videoID=videoID)
+        except:
+            return JsonResponse({'error': 4003, 'msg': "视频不存在"})
         liked_user = video.videoUpUser
-        LikeRecord.objects.get(likeUser=like_user, likedUser=liked_user, likeVideo=video).delete()
-        liked_user.TotalLikeNum = liked_user.TotalLikeNum - 1 # 更新用户点赞总数
-        liked_user.save()
-        video.videoLikeNum = video.videoLikeNum - 1
-        video.save()
-        return JsonResponse({'error': 0, 'msg': "取消点赞成功"})
+        if LikeRecord.objects.filter(likeUser=like_user, likedVideo=video).exists():
+            LikeRecord.objects.get(likeUser=like_user, likedUser=liked_user, likeVideo=video).delete()
+            liked_user.TotalLikeNum = liked_user.TotalLikeNum - 1 # 更新用户点赞总数
+            liked_user.save()
+            video.videoLikeNum = video.videoLikeNum - 1
+            video.save()
+            return JsonResponse({'error': 0, 'msg': "取消点赞成功"})
+        else:
+            return JsonResponse({'error': 4001, 'msg': "还未点赞"})
     else:
         return JsonResponse({'error': 2001, 'msg': "请求方式错误"})
 
@@ -52,12 +69,21 @@ def favourites(request):
     if request.method == 'POST':
         userID = request.POST.get('userID')
         videoID = request.POST.get('videoID')
-        user = UserInfo.objects.get(userID=userID)
-        video = VideoInfo.objects.get(videoID=videoID)
-        Favourites.objects.create(favorUser=user, favorVideo=video)
-        video.videoFavorNum = video.videoFavorNum + 1
-        video.save()
-        return JsonResponse({'error': 0, 'msg': "收藏成功"})
+        try:
+            user = UserInfo.objects.get(userID=userID)
+        except:
+            return JsonResponse({'error': 4002, 'msg': "用户不存在"})
+        try:
+            video = VideoInfo.objects.get(videoID=videoID)
+        except:
+            return JsonResponse({'error': 4003, 'msg': "视频不存在"})
+        if Favourites.objects.filter(favorUser=user, favorVideo=video).exists():
+            return JsonResponse({'error': 4001, 'msg': "已收藏"})
+        else:
+            Favourites.objects.create(favorUser=user, favorVideo=video)
+            video.videoFavorNum = video.videoFavorNum + 1
+            video.save()
+            return JsonResponse({'error': 0, 'msg': "收藏成功"})
     else:
         return JsonResponse({'error': 2001, 'msg': "请求方式错误"})
 
@@ -67,12 +93,21 @@ def cancalfavourites(request):
     if request.method == 'POST':
         userID = request.POST.get('userID')
         videoID = request.POST.get('videoID')
-        user = UserInfo.objects.get(userID=userID)
-        video = VideoInfo.objects.get(videoID=videoID)
-        Favourites.objects.get(favorUser=user, favorVideo=video).delete()
-        video.videoFavorNum = video.videoFavorNum - 1
-        video.save()
-        return JsonResponse({'error': 0, 'msg': "取消收藏成功"})
+        try:
+            user = UserInfo.objects.get(userID=userID)
+        except:
+            return JsonResponse({'error': 4002, 'msg': "用户不存在"})
+        try:
+            video = VideoInfo.objects.get(videoID=videoID)
+        except:
+            return JsonResponse({'error': 4003, 'msg': "视频不存在"})
+        if Favourites.objects.filter(favorUser=user, favorVideo=video).exists():
+            Favourites.objects.get(favorUser=user, favorVideo=video).delete()
+            video.videoFavorNum = video.videoFavorNum - 1
+            video.save()
+            return JsonResponse({'error': 0, 'msg': "取消收藏成功"})
+        else:
+            return JsonResponse({'error': 4001, 'msg': "还未收藏"})
     else:
         return JsonResponse({'error': 2001, 'msg': "请求方式错误"})
 
@@ -83,8 +118,14 @@ def comment(request):
         userID = request.POST.get('userID')
         videoID = request.POST.get('videoID')
         comment = request.POST.get('comment')
-        user = UserInfo.objects.get(userID=userID)
-        video = VideoInfo.objects.get(videoID=videoID)
+        try:
+            user = UserInfo.objects.get(userID=userID)
+        except:
+            return JsonResponse({'error': 4001, 'msg': "用户不存在"})
+        try:
+            video = VideoInfo.objects.get(videoID=videoID)
+        except:
+            return JsonResponse({'error': 4002, 'msg': "视频不存在"})
         commentted_user = video.videoUpUser
         VideoComment.objects.create(commentUpUser=commentted_user, commentComUser=user, commentVideo=video,
                                     commentContent=comment)
@@ -101,8 +142,14 @@ def cancelcomment(request):
         userID = request.POST.get('userID')
         videoID = request.POST.get('videoID')
         comment = request.POST.get('comment')
-        user = UserInfo.objects.get(userID=userID)
-        video = VideoInfo.objects.get(videoID=videoID)
+        try:
+            user = UserInfo.objects.get(userID=userID)
+        except:
+            return JsonResponse({'error': 4002, 'msg': "用户不存在"})
+        try:
+            video = VideoInfo.objects.get(videoID=videoID)
+        except:
+            return JsonResponse({'error': 4003, 'msg': "视频不存在"})
         commentted_user = video.videoUpUser
         try:
             VideoComment.objects.get(commentUpUser=commentted_user, commentComUser=user, commentVideo=video,
@@ -123,8 +170,14 @@ def editcomment(request):
         videoID = request.POST.get('videoID')
         comment = request.POST.get('comment')
         new_comment = request.POST.get('newComment')
-        user = UserInfo.objects.get(userID=userID)
-        video = VideoInfo.objects.get(videoID=videoID)
+        try:
+            user = UserInfo.objects.get(userID=userID)
+        except:
+            return JsonResponse({'error': 4002, 'msg': "用户不存在"})
+        try:
+            video = VideoInfo.objects.get(videoID=videoID)
+        except:
+            return JsonResponse({'error': 4003, 'msg': "视频不存在"})
         commentted_user = video.videoUpUser
         try:
             pre_comment = VideoComment.objects.get(commentUpUser=commentted_user, commentComUser=user, commentVideo=video,
@@ -144,7 +197,10 @@ def complaintvideo(request):
         videoID = request.POST.get('videoID')
         complaint_userID = request.POST.get('complaintUserID')
         complain_reason = request.POST.get('complainReason')
-        video = VideoInfo.objects.get(videoID=videoID)
+        try:
+            video = VideoInfo.objects.get(videoID=videoID)
+        except:
+            return JsonResponse({'error': 4001, 'msg': "视频不存在"})
         complaint_user = UserInfo.objects.get(userID=complaint_userID)
         complainted_user = video.videoUpUser
         AuditRecord.objects.create(auditVideo=video, complainUser=complaint_user, complainedUser=complainted_user,
